Address EX31 review feedback locally
