+++ 31_HLSLPathTracer/app_resources/hlsl/compute.render.linear.entrypoints.hlsl
@@ -3,7 +3,7 @@
 #endif
 
 #ifndef PATH_TRACER_ENTRYPOINT_POLYGON_METHOD
-#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PPM_APPROX_PROJECTED_SOLID_ANGLE
+#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD pathtracer_variant_config::EntryPointPolygonMethod
 #endif
 
 #if !PATH_TRACER_ENABLE_LINEAR
