+++ 31_HLSLPathTracer/app_resources/hlsl/compute_render_scene_impl.hlsl
@@ -156,4 +156,3 @@ void runPersistent()
 }
 #endif
 }
-#undef PATH_TRACER_USE_RWMC
