+++ 31_HLSLPathTracer/app_resources/hlsl/spirv/pt.compute.triangle.methods.shared.hlsl
@@ -6,7 +6,6 @@
 
 #define PT_VARIANT_ENTRYPOINT_LINEAR 1
 #define PT_VARIANT_ENTRYPOINT_PERSISTENT 2
-#define PATH_TRACER_USE_RWMC PT_VARIANT_USE_RWMC
 #if PT_VARIANT_ENTRYPOINT_KIND == PT_VARIANT_ENTRYPOINT_LINEAR
 #define PATH_TRACER_ENABLE_LINEAR 1
 #define PATH_TRACER_ENABLE_PERSISTENT 0
@@ -17,7 +16,21 @@
 #error Unsupported PT_VARIANT_ENTRYPOINT_KIND
 #endif
 
+#if PT_VARIANT_USE_RWMC
+#define PATH_TRACER_VARIANT_USE_RWMC true
+#else
+#define PATH_TRACER_VARIANT_USE_RWMC false
+#endif
+
+#define PATH_TRACER_VARIANT_GEOMETRY ELG_TRIANGLE
+#define PATH_TRACER_VARIANT_POLYGON_METHOD EPM_PROJECTED_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD PPM_APPROX_PROJECTED_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_ENABLE_LINEAR PATH_TRACER_ENABLE_LINEAR
+#define PATH_TRACER_VARIANT_ENABLE_PERSISTENT PATH_TRACER_ENABLE_PERSISTENT
+#define PATH_TRACER_USE_RWMC PT_VARIANT_USE_RWMC
+
 #include "compute.render.common.hlsl"
+#include "nbl/this_example/render_variant_config.hlsl"
 #include "scene_triangle_light.hlsl"
 #include "compute_render_scene_impl.hlsl"
 
@@ -48,3 +61,11 @@
 #define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
 #include "compute.render.persistent.entrypoints.hlsl"
 #endif
+
+#undef PATH_TRACER_VARIANT_ENABLE_PERSISTENT
+#undef PATH_TRACER_VARIANT_ENABLE_LINEAR
+#undef PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD
+#undef PATH_TRACER_VARIANT_POLYGON_METHOD
+#undef PATH_TRACER_VARIANT_GEOMETRY
+#undef PATH_TRACER_VARIANT_USE_RWMC
+#undef PATH_TRACER_USE_RWMC
