+++ 31_HLSLPathTracer/app_resources/hlsl/spirv/pt.compute.variant.shared.hlsl
@@ -10,7 +10,6 @@
 #define PT_VARIANT_ENTRYPOINT_LINEAR 1
 #define PT_VARIANT_ENTRYPOINT_PERSISTENT 2
 
-#define PATH_TRACER_USE_RWMC PT_VARIANT_USE_RWMC
 #if PT_VARIANT_ENTRYPOINT_KIND == PT_VARIANT_ENTRYPOINT_COMBINED
 #define PATH_TRACER_ENABLE_LINEAR 1
 #define PATH_TRACER_ENABLE_PERSISTENT 1
@@ -24,29 +23,50 @@
 #error Unsupported PT_VARIANT_ENTRYPOINT_KIND
 #endif
 
+#if PT_VARIANT_USE_RWMC
+#define PATH_TRACER_VARIANT_USE_RWMC true
+#else
+#define PATH_TRACER_VARIANT_USE_RWMC false
+#endif
+
+#define PATH_TRACER_VARIANT_ENABLE_LINEAR PATH_TRACER_ENABLE_LINEAR
+#define PATH_TRACER_VARIANT_ENABLE_PERSISTENT PATH_TRACER_ENABLE_PERSISTENT
+#define PATH_TRACER_USE_RWMC PT_VARIANT_USE_RWMC
+
 #include "compute.render.common.hlsl"
 #if PT_VARIANT_SCENE_KIND == PT_VARIANT_SCENE_SPHERE
-#define PT_VARIANT_POLYGON_METHOD PPM_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_GEOMETRY ELG_SPHERE
+#define PATH_TRACER_VARIANT_POLYGON_METHOD EPM_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
+#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
 #include "scene_sphere_light.hlsl"
 #elif PT_VARIANT_SCENE_KIND == PT_VARIANT_SCENE_RECTANGLE
-#define PT_VARIANT_POLYGON_METHOD PPM_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_GEOMETRY ELG_RECTANGLE
+#define PATH_TRACER_VARIANT_POLYGON_METHOD EPM_SOLID_ANGLE
+#define PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
+#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
 #include "scene_rectangle_light.hlsl"
 #else
 #error Unsupported PT_VARIANT_SCENE_KIND
 #endif
+#include "nbl/this_example/render_variant_config.hlsl"
 #include "compute_render_scene_impl.hlsl"
 
 #if PT_VARIANT_ENTRYPOINT_KIND == PT_VARIANT_ENTRYPOINT_COMBINED
-#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PT_VARIANT_POLYGON_METHOD
 #include "compute.render.linear.entrypoints.hlsl"
-#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PT_VARIANT_POLYGON_METHOD
+#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PPM_SOLID_ANGLE
 #include "compute.render.persistent.entrypoints.hlsl"
 #elif PT_VARIANT_ENTRYPOINT_KIND == PT_VARIANT_ENTRYPOINT_LINEAR
-#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PT_VARIANT_POLYGON_METHOD
 #include "compute.render.linear.entrypoints.hlsl"
 #else
-#define PATH_TRACER_ENTRYPOINT_POLYGON_METHOD PT_VARIANT_POLYGON_METHOD
 #include "compute.render.persistent.entrypoints.hlsl"
 #endif
 
-#undef PT_VARIANT_POLYGON_METHOD
+#undef PATH_TRACER_ENTRYPOINT_POLYGON_METHOD
+#undef PATH_TRACER_USE_RWMC
+#undef PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD
+#undef PATH_TRACER_VARIANT_POLYGON_METHOD
+#undef PATH_TRACER_VARIANT_GEOMETRY
+#undef PATH_TRACER_VARIANT_ENABLE_PERSISTENT
+#undef PATH_TRACER_VARIANT_ENABLE_LINEAR
+#undef PATH_TRACER_VARIANT_USE_RWMC
