+++ 31_HLSLPathTracer/include/nbl/this_example/render_variant_config.hlsl
@@ -1,0 +1,40 @@
+#ifndef __NBL_THIS_EXAMPLE_RENDER_VARIANT_CONFIG_HLSL_INCLUDED__
+#define __NBL_THIS_EXAMPLE_RENDER_VARIANT_CONFIG_HLSL_INCLUDED__
+
+#include "nbl/this_example/render_variant_enums.hlsl"
+
+#ifndef PATH_TRACER_VARIANT_GEOMETRY
+#error PATH_TRACER_VARIANT_GEOMETRY must be defined before including render_variant_config.hlsl
+#endif
+
+#ifndef PATH_TRACER_VARIANT_POLYGON_METHOD
+#error PATH_TRACER_VARIANT_POLYGON_METHOD must be defined before including render_variant_config.hlsl
+#endif
+
+#ifndef PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD
+#error PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD must be defined before including render_variant_config.hlsl
+#endif
+
+#ifndef PATH_TRACER_VARIANT_USE_RWMC
+#error PATH_TRACER_VARIANT_USE_RWMC must be defined before including render_variant_config.hlsl
+#endif
+
+#ifndef PATH_TRACER_VARIANT_ENABLE_LINEAR
+#error PATH_TRACER_VARIANT_ENABLE_LINEAR must be defined before including render_variant_config.hlsl
+#endif
+
+#ifndef PATH_TRACER_VARIANT_ENABLE_PERSISTENT
+#error PATH_TRACER_VARIANT_ENABLE_PERSISTENT must be defined before including render_variant_config.hlsl
+#endif
+
+struct pathtracer_variant_config
+{
+	NBL_CONSTEXPR_STATIC_INLINE E_LIGHT_GEOMETRY Geometry = PATH_TRACER_VARIANT_GEOMETRY;
+	NBL_CONSTEXPR_STATIC_INLINE E_POLYGON_METHOD PolygonMethod = PATH_TRACER_VARIANT_POLYGON_METHOD;
+	NBL_CONSTEXPR_STATIC_INLINE NEEPolygonMethod EntryPointPolygonMethod = PATH_TRACER_VARIANT_ENTRYPOINT_POLYGON_METHOD;
+	NBL_CONSTEXPR_STATIC_INLINE bool UseRWMC = PATH_TRACER_VARIANT_USE_RWMC;
+	NBL_CONSTEXPR_STATIC_INLINE bool EnableLinear = PATH_TRACER_VARIANT_ENABLE_LINEAR;
+	NBL_CONSTEXPR_STATIC_INLINE bool EnablePersistent = PATH_TRACER_VARIANT_ENABLE_PERSISTENT;
+};
+
+#endif
