+++ 31_HLSLPathTracer/include/nbl/this_example/render_variant_enums.hlsl
@@ -1,7 +1,11 @@
 #ifndef __NBL_THIS_EXAMPLE_RENDER_VARIANT_ENUMS_HLSL_INCLUDED__
 #define __NBL_THIS_EXAMPLE_RENDER_VARIANT_ENUMS_HLSL_INCLUDED__
 
+#ifndef __HLSL_VERSION
 enum E_LIGHT_GEOMETRY : uint8_t
+#else
+enum E_LIGHT_GEOMETRY
+#endif
 {
 	ELG_SPHERE,
 	ELG_TRIANGLE,
@@ -9,7 +13,11 @@ enum E_LIGHT_GEOMETRY : uint8_t
 	ELG_COUNT
 };
 
+#ifndef __HLSL_VERSION
 enum E_POLYGON_METHOD : uint8_t
+#else
+enum E_POLYGON_METHOD
+#endif
 {
 	EPM_AREA,
 	EPM_SOLID_ANGLE,
